Fix warnings

+++ module/powerflow/solver_py.cpp
@@ -921,7 +921,7 @@ void sync_busdata_raw(PyObject *pModel,unsigned int &bus_count,BUSDATA *&bus,e_d
 			"IAr","IAi","IBr","IBi","ICr","ICi",
 			"VAm","VAa","VBm","VBa","VCm","VCa",
 		};
-		size_t ntags = sizeof(tags)/sizeof(tags[0]);
+		long int ntags = sizeof(tags)/sizeof(tags[0]);
 		PyObject *taglist = PyList_New(ntags);
 		for ( size_t m = 0 ; m < ntags ; m++ )
 		{
